requirements.txt
django==4.1.7
psycopg2-binary==2.9.5
python-dotenv==0.21.1
django-environ==0.9.0
djangorestframework==3.14.0
drf-yasg==1.21.5
pillow==9.4.0
whitenoise==6.3.0
djangorestframework-simplejwt==5.2.2